plat: marvell: octeontx: Enable MRVL_TF_LOGGING
Enable MRVL_TF_LOGGING and fix compilation errors.
Use long instead of long long on aarch64 for 64_t stdint types.

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/95488
Tested-by: Chandrakala Chavva <cchavva@marvell.com>

+++ plat/marvell/octeontx/octeontx_common.c
@@ -405,7 +405,7 @@ int32_t plat_get_soc_revision(void)
 }
 
 #ifdef MRVL_TF_LOG_MODULE
-unsigned int mrvl_tf_log_modules;
+unsigned long mrvl_tf_log_modules;
 
 /*
  * initialize_tf_logging()
